+++ store/plasmaStore.go
@@ -3,6 +3,7 @@ package store
 import (
 	"fmt"
 	"github.com/FourthState/plasma-mvp-sidechain/plasma"
+	"github.com/FourthState/plasma-mvp-sidechain/utils"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/rlp"
 	"math/big"
@@ -13,9 +14,10 @@ type PlasmaStore struct {
 }
 
 const (
-	confirmSigPrefix = "confirmSignature"
-	deposit          = "deposit"
-	blockKey         = "block"
+	confirmSigKey     = "confirmSignature"
+	blockKey          = "block"
+	plasmaBlockNumKey = "plasmaBlockNum"
+	plasmaToTmKey     = "plasmatotm"
 )
 
 func NewPlasmaStore(ctxKey sdk.StoreKey) PlasmaStore {
@@ -25,7 +27,7 @@ func NewPlasmaStore(ctxKey sdk.StoreKey) PlasmaStore {
 }
 
 func (store PlasmaStore) GetBlock(ctx sdk.Context, blockHeight *big.Int) (plasma.Block, bool) {
-	key := CreateBlockKey(blockHeight)
+	key := prefixKey(blockKey, blockHeight.Bytes())
 	data := store.Get(ctx, key)
 	if data == nil {
 		return plasma.Block{}, false
@@ -39,18 +41,30 @@ func (store PlasmaStore) GetBlock(ctx sdk.Context, blockHeight *big.Int) (plasma
 	return block, true
 }
 
-func (store PlasmaStore) StoreBlock(ctx sdk.Context, blockHeight *big.Int, block plasma.Block) {
-	key := CreateBlockKey(blockHeight)
-	data, err := rlp.EncodeToBytes(&block)
+// StoreBlock will store the plasma block and return the plasma block number in which it was stored under
+func (store PlasmaStore) StoreBlock(ctx sdk.Context, tmBlockHeight *big.Int, block plasma.Block) *big.Int {
+	plasmaBlockNum := store.NextPlasmaBlockNum(ctx)
+
+	plasmaBlockKey := prefixKey(blockKey, plasmaBlockNum.Bytes())
+	plasmaBlockData, err := rlp.EncodeToBytes(&block)
 	if err != nil {
 		panic(fmt.Sprintf("error rlp encoding block: %s", err))
 	}
 
-	store.Set(ctx, key, data)
+	// store the block
+	store.Set(ctx, plasmaBlockKey, plasmaBlockData)
+
+	// latest plasma block number
+	store.Set(ctx, []byte(plasmaBlockNumKey), plasmaBlockNum.Bytes())
+
+	// plasma block number => tendermint block numper
+	store.Set(ctx, prefixKey(plasmaToTmKey, plasmaBlockNum.Bytes()), tmBlockHeight.Bytes())
+
+	return plasmaBlockNum
 }
 
 func (store PlasmaStore) StoreConfirmSignatures(ctx sdk.Context, position plasma.Position, confirmSignatures [][65]byte) {
-	key := CreateConfirmSignatureKey(position)
+	key := prefixKey(confirmSigKey, position.Bytes())
 
 	var sigs []byte
 	sigs = append(sigs, confirmSignatures[0][:]...)
@@ -61,13 +75,17 @@ func (store PlasmaStore) StoreConfirmSignatures(ctx sdk.Context, position plasma
 	store.Set(ctx, key, sigs)
 }
 
-func CreateBlockKey(blockHeight *big.Int) []byte {
-	return prefixKey(blockKey, blockHeight.Bytes())
-}
+func (store PlasmaStore) NextPlasmaBlockNum(ctx sdk.Context) *big.Int {
+	var plasmaBlockNum *big.Int
+	data := store.Get(ctx, []byte(plasmaBlockNumKey))
+	if data == nil {
+		plasmaBlockNum = big.NewInt(1)
+	} else {
+		plasmaBlockNum = new(big.Int).SetBytes(data)
 
-func CreateConfirmSignatureKey(pos plasma.Position) []byte {
-	// zero out the output index to prevent duplicates
-	pos.OutputIndex = 0
+		// increment the block number
+		plasmaBlockNum = plasmaBlockNum.Add(plasmaBlockNum, utils.Big1)
+	}
 
-	return prefixKey(confirmSigPrefix, pos.Bytes())
+	return plasmaBlockNum
 }
